chore(RaceProgressFormatter): relocate class and test from domain to formatter package for clearer responsibility separation
