+++ src/main/kotlin/racingcar/formatter/RaceProgressFormatter.kt
@@ -1,4 +1,6 @@
-package racingcar.domain
+package racingcar.formatter
+
+import racingcar.domain.Cars
 
 class RaceProgressFormatter {
     fun format(cars: Cars): List<String> {
