+++ src/test/kotlin/racingcar/formatter/RaceProgressFormatterTest.kt
@@ -1,7 +1,9 @@
-package racingcar.domain
+package racingcar.formatter
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.domain.CarName
+import racingcar.domain.Cars
 
 class RaceProgressFormatterTest {
